requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
gspread==3.6.0
httplib2==0.18.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
iso-639==0.4.5
iso3166==1.0.1
isodate==0.6.0
moviepy==1.0.3
numpy==1.19.2
pandas==1.1.3
Pillow==7.2.0
proglog==0.1.9
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.8
PySocks==1.7.1
python-dateutil==2.8.1
pytube3 @ git+https://gitlab.com/obuilds/public/pytube@10c57109f87fe864d8f38bbc8d76941e695de93a
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
streamlink==1.6.0
tqdm==4.50.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.10
websocket-client==0.57.0
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
requests==2.24.0
requests-oauthlib==1.3.0
appdirs==1.4.4
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
distlib==0.3.1
filelock==3.0.12
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
requests==2.24.0
requests-oauthlib==1.3.0
httplib2==0.18.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
moviepy==1.0.3
numpy==1.19.2
oauthlib==3.1.0
Pillow==7.2.0
proglog==0.1.9
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
rsa==4.6
six==1.15.0
tqdm==4.50.2
uritemplate==3.0.1
urllib3==1.25.10
appdirs==1.4.4
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
distlib==0.3.1
ffmpeg==1.4
filelock==3.0.12
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
gspread==3.6.0
httplib2==0.18.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
iso-639==0.4.5
iso3166==1.0.1
isodate==0.6.0
moviepy==1.0.3
numpy==1.19.2
oauth2client==4.1.3
oauthlib==3.1.0
pandas==1.1.3
Pillow==7.2.0
proglog==0.1.9
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.8
PySocks==1.7.1
python-dateutil==2.8.1
pytube3 @ git+https://gitlab.com/obuilds/public/pytube@10c57109f87fe864d8f38bbc8d76941e695de93a
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
streamlink==1.6.0
tqdm==4.50.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.10
websocket-client==0.57.0
youtube-dl==2020.9.20

